dependabot: Disable k8s minor updates
Each minor OpenShift version uses the same minor k8s version throughout
its lifetime. This minor k8s version used by OpenShift can be behind the latest upstream minor version.
This  commit changes the dependabot configuration so that it does not
automatically attempt to update the minor k8s version, as it's better if
we use the same one as the one used by the cluster.

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>